Added gui images folder for the new gui coming, and began working on a simple GUIHandler
The GUI works by first loading all the images and storing it in the ImageLoader (see ImageLoader class), then the GUIHandler contains references to all of the GUIObjects. The GUIHandler updates and draws each of these GUIObjects automatically.
To add a GUIObject to the GUIHandler and therefore into the game, go into the setup method (void setup()) and see the example there